add publish workflow, reorg files
